changed encode interfaces
