+++ encode/flag/unload.go
@@ -1,0 +1,13 @@
+package flg
+
+import "bytes"
+
+func NewUnloader() *Unloader {
+	return &Unloader{
+		buf: &bytes.Buffer{},
+	}
+}
+
+type Unloader struct {
+	buf *bytes.Buffer
+}
